Refactor and typo fixes from code review
